requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
eventlet==0.35.2
greenlet==3.2.4
gunicorn==22.0.0
SQLAlchemy==2.0.43
python-dotenv==1.0.1
itsdangerous==2.2.0
Werkzeug==3.0.3
click==8.2.1
blinker==1.9.0
dnspython==2.7.0
Flask-Mail==0.10.0
email-validator==2.3.0
pydantic==2.8.2
tenacity==9.1.2
requests==2.32.5
playwright
redis
rq
python-telegram-bot==13.15
psycopg2-binary
celery

